Add ssh files for cnx-recipes fetch-html script
fetch-html uses ssh to download a book in single html format.  Add a ssh
config file to accept new host keys automatically in order to suppress
the prompt to add new host keys to the ssh known hosts file.  If
fetch_book_rsa exists, it is used as the ssh key in fetch-html.