Update README with more ssh key set up for fetch-html
Since fetch-html takes some time to complete and there are 2 password
prompts if ssh key is not set up, it is best to set up an ssh key so
that the user can run the script without password prompts.

Also, include examples of how to use environment variables when running
`fetch-html` and `bake`.